SCMNEW-4: service provider

+++ scm-common-data/src/main/java/ir/daneshrefah/scm/common/data/entity/asset/MembershipTerminalServiceAccessEntity.java
@@ -25,5 +25,5 @@ public class MembershipTerminalServiceAccessEntity extends AbstractEntity<Intege
     @JoinColumn(name = "MCS_ID")
     private MembershipTerminalAccessEntity membershipTerminalAccess;
     private Integer archiveNo;
-    private BigDecimal maxWithdrawalPerTransaction;
+    private Long maxWithdrawalPerTransaction;
 }
